refactor(stac)!: remove io module
